backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
groq==0.4.1
web3==6.12.0
eth-account==0.9.0
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0
structlog==23.2.0
tenacity==8.2.3
aiofiles==23.2.1
jinja2==3.1.2
email-validator==2.1.0
